Dodanie opcji przypomnienia hasła oraz jego zmiany

+++ Projekt_PO_w61933/ShowPackage.xaml.cs
@@ -38,36 +38,40 @@ namespace Projekt_PO_w61933
 
         private void bDeletePackage_Click(object sender, RoutedEventArgs e)
         {
-            if (lbAllPackage.SelectedItem.ToString() != "")
+            if (lbAllPackage.SelectedIndex > -1)
             {
-                MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz wyłączyć ten pakiet?", "",
-                                 MessageBoxButton.YesNo,
-                                 MessageBoxImage.Question);
-                if (lbAllPackage.SelectedIndex > 0 && result == MessageBoxResult.Yes)
+                if (lbAllPackage.SelectedItem.ToString() != "")
                 {
+                    MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz wyłączyć ten pakiet?", "",
+                                     MessageBoxButton.YesNo,
+                                     MessageBoxImage.Question);
+                    if (lbAllPackage.SelectedIndex > 0 && result == MessageBoxResult.Yes)
+                    {
 
-                    string selectedPackage = lbAllPackage.SelectedItem.ToString();
-                    string packageState = File.ReadLines("Packages.txt").Skip(id - 1).Take(1).First();
-                    string oldPackageState = packageState;
-                    packageState = packageState.Replace(selectedPackage + ';', "");
-                    string packagestxt = File.ReadAllText("Packages.txt");
+                        string selectedPackage = lbAllPackage.SelectedItem.ToString();
+                        string packageState = File.ReadLines("Packages.txt").Skip(id - 1).Take(1).First();
+                        string oldPackageState = packageState;
+                        packageState = packageState.Replace(selectedPackage + ';', "");
+                        string packagestxt = File.ReadAllText("Packages.txt");
 
-                    packagestxt = packagestxt.Replace(oldPackageState, packageState);
+                        packagestxt = packagestxt.Replace(oldPackageState, packageState);
 
-                    File.WriteAllText("Packages.txt", packagestxt);
-                    MessageBox.Show("Wyłączyłeś pakiet: "+selectedPackage);
-                    this.DialogResult = true;
-                    MainWindow mainWindow = new MainWindow();
-                    mainWindow.showUserInterface(id);
-                    mainWindow.Close();
-                }
-                else if (lbAllPackage.SelectedIndex == -1)
-                {
-                    MessageBox.Show("Nie wybrałeś pakietu");
+                        File.WriteAllText("Packages.txt", packagestxt);
+                        MessageBox.Show("Wyłączyłeś pakiet: " + selectedPackage);
+                        OperationsUser operationsUser = new OperationsUser("Wyłączenie pakietu:", selectedPackage, id);
+                        this.DialogResult = true;
+                        MainWindow mainWindow = new MainWindow();
+                        mainWindow.showUserInterface(id);
+                        mainWindow.Close();
+                    }
+                    
                 }
             }
-            
-            
+            else if (lbAllPackage.SelectedIndex == -1)
+            {
+                MessageBox.Show("Nie wybrałeś pakietu");
+            }
+
         }
     }
 }
